Widget fixes, namespace fixed, some util scripts moved inside DLL, cleanup, getAdvertisingId fix
