--- Assets/Odnoklassniki/Scripts/AbstractOdnoklassniki.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9f3c4091bafd6d345bea0ff8e74d7823
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
