--- Assets/Odnoklassniki/Scripts/AndroidOdnoklassniki.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 137e1cf07e35e274bb20ab56811afd1f
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
