--- Assets/Odnoklassniki/Scripts/IOSOdnoklassniki.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c6d8900701ee3bf409f2020ad9d2a364
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
