--- Assets/Odnoklassniki/Scripts/IOdnoklassniki.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 35702e02576b5184ba16d1940fcfbc4f
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
