--- Assets/Odnoklassniki/Scripts/OK.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d78e77c9c05f5e845b90bf0e3581c2d0
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
