--- Assets/Odnoklassniki/Scripts/OKSettings.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1af17cdb4fb467b4f8d6783aab80d7ae
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
