--- Assets/Odnoklassniki/Scripts/Util.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 122f485e5dbc7de449fe964d78ac64ff
-folderAsset: yes
-DefaultImporter:
-  userData:
