--- Assets/Odnoklassniki/Scripts/Util/Android.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: eeaac815f3927be47af6c92b9f0787e5
-folderAsset: yes
-DefaultImporter:
-  userData:
