--- Assets/Odnoklassniki/Scripts/Util/Android/ManifestMod.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0800e57767b4fcc4ca0c0ed93d4da51d
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
