--- Assets/Odnoklassniki/Scripts/Util/Android/OdnoklassnikiAndroidUtil.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fea86be92f353b1488ceaaf4a54e4df5
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
