--- Assets/Odnoklassniki/Scripts/Util/Encryption.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 19d6462a502398e4d929acb73410b74d
-folderAsset: yes
-DefaultImporter:
-  userData:
