--- Assets/Odnoklassniki/Scripts/Util/Encryption/Encryption.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dd15f7d49df02684bb55f3117ebaa02d
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
