--- Assets/Odnoklassniki/Scripts/Util/OKWebView.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 246e81e63e7ab964e94e65b4a9462caa
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
