--- Assets/Odnoklassniki/Scripts/Util/iOS.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 8024cc5eba1d18d478892a476aee9d90
-folderAsset: yes
-DefaultImporter:
-  userData:
