--- Assets/Odnoklassniki/Scripts/Util/iOS/PListParser.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 40c7c8facdb40b34e859d20bd2ce60e2
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
