--- Assets/Odnoklassniki/Scripts/Util/iOS/PListWrapper.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3282d590e2cd34442b67b0535880d3f6
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
