+++ Assets/Plugins/OKWebViewPluginAndroid.cs
@@ -1,44 +1,47 @@
 ﻿#if UNITY_ANDROID
 using UnityEngine;
 
-public class OKWebViewPlugin {
-
-	private static AndroidJavaClass java;
-
-	public static void Init(string name)
+namespace Odnoklassniki.WebView
+{
+	public class OKWebViewPlugin
 	{
-		java = new AndroidJavaClass("ru.odnoklassniki.unity.OKAndroidPlugin");
-		java.CallStatic("OKWV_Init", name);
-	}
+		private static AndroidJavaClass java;
 
-	public static void Load(string url)
-	{
-		java.CallStatic("OKWV_Load", url);
-	}
+		public static void Init(string name)
+		{
+			java = new AndroidJavaClass("ru.odnoklassniki.unity.OKAndroidPlugin");
+			java.CallStatic("OKWV_Init", name);
+		}
 
-	public static void Show()
-	{
-		java.CallStatic("OKWV_Show");
-	}
+		public static void Load(string url)
+		{
+			java.CallStatic("OKWV_Load", url);
+		}
 
-	public static void Hide()
-	{
-		java.CallStatic("OKWV_Hide");
-	}
+		public static void Show()
+		{
+			java.CallStatic("OKWV_Show");
+		}
 
-	public static void ClearCookies()
-	{
-		java.CallStatic("OKWV_ClearCookies");
-	}
+		public static void Hide()
+		{
+			java.CallStatic("OKWV_Hide");
+		}
 
-	public static void Destroy()
-	{
-		java.CallStatic("OKWV_Destroy");
-	}
+		public static void ClearCookies()
+		{
+			java.CallStatic("OKWV_ClearCookies");
+		}
 
-	public static void Resize()
-	{
-		java.CallStatic("OKWV_Resize");
+		public static void Destroy()
+		{
+			java.CallStatic("OKWV_Destroy");
+		}
+
+		public static void Resize()
+		{
+			java.CallStatic("OKWV_Resize");
+		}
 	}
 }
 #endif
